# backend/api/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AapadBandhav Smart Emergency Response API",
    "version": "2.0.0",
    "description": "## AapadBandhav Platform API\n\n**AapadBandhav** is a Smart Emergency Response and Accident Detection Platform for India.\n\n### Features\n- 🚨 Real-time accident detection via IoT crash sensors\n- 🚑 Smart dispatch to hospitals, ambulances, police, fire departments\n- 🗺️ Live GPS tracking of responders and affected users\n- 📲 OTP-based authentication for all entity types\n- 🛡️ Full RBAC with 11 roles\n\n### Authentication\nAll protected endpoints require a **Bearer JWT token**.\n\nTo authenticate:\n1. Send OTP → `POST /api/auth/otp/send`\n2. Verify OTP → `POST /api/auth/otp/verify` → get **token**\n3. Click **Authorize** above → enter: `Bearer <your-token>`\n\nFor admin access use `POST /api/auth/admin/login`.\n\n### Roles\n`user` · `volunteer` · `fire_department` · `emergency_personnel`  \n`hospital` · `ambulance` · `police_station` · `policeman`  \n`mechanic` · `insurance` · `admin` · `superadmin`",
    "contact": {
      "name": "AapadBandhav Team",
      "email": "admin@aapadbandhav.in"
    },
    "license": {
      "name": "Private"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    },
    {
      "url": "https://aapadbandhav.vercel.app",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token. Get one from POST /api/auth/otp/verify or POST /api/auth/admin/login"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occurred"
          }
        }
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "example": "Accident Alert"
          },
          "body": {
            "type": "string",
            "example": "An accident has been detected near your location."
          },
          "type": {
            "type": "string",
            "example": "accident_alert"
          },
          "data": {
            "type": "object"
          },
          "read": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "security": [],
...
  "paths": {
...
    "/api/iot/ingest": {
      "post": {
        "tags": [
          "IoT"
        ],
        "summary": "Receive IoT device telemetry (MQTT over HTTP bridge)",
        "description": "Receives MQTT-format telemetry from IoT crash sensors via HTTP bridge.\nProcesses GPS position, speed, impact G-force readings, and battery status.\n\n**Auto-Crash Detection:** If `impactValue >= 3.0G`, automatically triggers an accident\nand initiates the full emergency dispatch pipeline (Inngest workflow, Pusher broadcast,\nalert dispatch to nearby responders).\n\n**Topic Format:** `vehicle/{deviceId}/{nodeId}`\n- `nodeId` identifies the physical sensor (front, rear, side)\n\n**Authentication:** No JWT required — authenticated via device credentials embedded in the topic/payload.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "topic",
                  "payload"
                ],
                "properties": {
                  "topic": {
                    "type": "string",
                    "example": "vehicle/4810881048888104/front",
                    "description": "MQTT topic in format vehicle/{deviceId}/{nodeId}"
                  },
                  "payload": {
                    "type": "object",
                    "description": "Sensor telemetry data",
                    "properties": {
                      "latitude": {
                        "type": "number",
                        "example": 16.5062
                      },
                      "longitude": {
                        "type": "number",
                        "example": 80.648
                      },
                      "speed": {
                        "type": "number",
                        "example": 72.5,
                        "description": "Speed in km/h"
                      },
                      "impactValue": {
                        "type": "number",
                        "example": 8.4,
                        "description": "Crash impact in G-force"
                      },
                      "batteryStatus": {
                        "type": "number",
                        "example": 85.5,
                        "description": "Battery percentage"
                      },
                      "sensorStatus": {
                        "type": "string",
                        "example": "active"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Telemetry processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing topic",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get my notifications",
        "description": "Returns the last 50 in-app notifications for the authenticated entity.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notification list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "notifications": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notification"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/notifications/fcm-token": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Register FCM push notification token",
        "description": "Stores the Firebase Cloud Messaging (FCM) device token for the authenticated entity, enabling push notifications on accident and dispatch events.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fcm_token"
                ],
                "properties": {
                  "fcm_token": {
                    "type": "string",
                    "example": "cNHj4dMoQ-GhV5kTkZY9xE:APA91bHqz...",
                    "description": "Firebase Cloud Messaging device token"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "FCM token registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}